After a year in use, tweak parts
Took the van apart ready to do the insulation and flooring so
I have all of the parts in the house for the first time in ages.
Taking the opportunity to tweak the design to make it better and
also switch from using screws in the bed frams to clevis pins
making this easier to keep secure but remove when needed.

* Make the top taller
* Narrow the whole thing to fit between the supports
* Switch from two fixing holes to one on top
* Increase the screw hole size to better support 4mm clevis pins